warning message asking people for timestamp
I came to realise that many new transcriptions are started without any information, how far the person came doing the transcription. 

This is not bad, if only a few lines were added and you see someone just tried it out, but it takes quite some time to find the end of a quite a long transcription in a long video. Also we would of course like to have better statistics!