Fix m2m talk.persons migration
Django is not able to perform AlterField migrations that change
ManyToMany fields:

ValueError: Cannot alter field %s into %s - they are not compatible
types (you cannot alter to or from M2M fields, or add or remove
through= on M2M fields)

A workaround is to pretend to create a new model for the relationship
table and remove/add the field via the SeparateDatabaseAndState
migration. A SQL migration is not needed, because no schema changes are
required.